refactor(guard): isolate false-positive precision rules
