+++ ci/code-quality-baseline.json
@@ -515,7 +515,7 @@
     "src/codex_plugin_scanner/guard/cli/update_commands.py": 2838,
     "src/codex_plugin_scanner/guard/cli/update_subprocess.py": 2087,
     "src/codex_plugin_scanner/guard/codex_app_server.py": 627,
-    "src/codex_plugin_scanner/guard/config.py": 1392,
+    "src/codex_plugin_scanner/guard/config.py": 1396,
     "src/codex_plugin_scanner/guard/consumer/service.py": 1859,
     "src/codex_plugin_scanner/guard/daemon/dashboard_update.py": 667,
     "src/codex_plugin_scanner/guard/daemon/extension_control_api.py": 509,
