+++ ci/test-suite-ratchet-baseline.json
@@ -1,5 +1,5 @@
 {
   "maximum_collected_cases": 15397,
-  "maximum_test_source_lines": 332143,
+  "maximum_test_source_lines": 332151,
   "schema_version": 1
 }
