+++ src/codex_plugin_scanner/guard/risk.py
@@ -8,6 +8,7 @@ from pathlib import PurePath
 from urllib.parse import urlsplit
 
 from .models import GuardArtifact
+from .risk_exfil import detect_exfil_intent
 from .runtime.secret_sensitivity import classify_legacy_secret_path_families
 from .runtime.signals import RiskSignalV2
 from .types import GuardSignal
@@ -61,23 +62,6 @@ _GUARD_BYPASS_PATTERNS: tuple[tuple[str, str], ...] = (
     (r"approval_policy\s*=\s*\"never\"", "approval policy forced to never"),
     (r"\.codex/config\.toml", "direct Guard-managed configuration mutation"),
     (r"guard[_-]?bypass", "explicit Guard bypass marker"),
-)
-_EXFIL_PATTERNS: tuple[tuple[str, str], ...] = (
-    (r"\bexfiltrat(?:e|es|ed|ing|ion)\b", "explicit exfiltration intent"),
-    (r"(gist\.github\.com|pastebin\.com|transfer\.sh|webhook)", "external sink destination"),
-    (r"scp\s+", "scp transfer intent"),
-)
-_EXFIL_TRANSFER_VERB_PATTERN = re.compile(r"\b(?:upload|send|post|sync)\b")
-_EXFIL_SECRET_SOURCE_PATTERN = re.compile(
-    "".join(
-        (
-            r"(?:\.env\b|\.npmrc\b|\.pypirc\b|\.ssh/|\.aws/credentials\b|\.kube/config\b|",
-            r"\b(?:api[_-]?key|auth[_-]?token|access[_-]?token|credential|password|private[_-]?key|secret)\b)",
-        )
-    )
-)
-_EXFIL_NETWORK_SINK_PATTERN = re.compile(
-    r"(?:https?://|\b(?:curl|wget|scp)\b|gist\.github\.com|pastebin\.com|transfer\.sh|webhook)"
 )
 _SECRET_PATH_LABELS: tuple[tuple[str, str], ...] = (
     (".env", "local .env file"),
@@ -398,35 +382,6 @@ def detect_guard_bypass(text: str) -> list[GuardSignal]:
                     matched_text=reason,
                     explanation="contains guard bypass intent",
                     remediation="Block and require manual investigation.",
-                    rule_version=_RULE_VERSION,
-                )
-            )
-    return signals
-
-
-def detect_exfil_intent(text: str) -> list[GuardSignal]:
-    """Detect exfiltration-oriented phrasing and destinations."""
-
-    signals: list[GuardSignal] = []
-    patterns = list(_EXFIL_PATTERNS)
-    if (
-        _EXFIL_TRANSFER_VERB_PATTERN.search(text)
-        and _EXFIL_SECRET_SOURCE_PATTERN.search(text)
-        and _EXFIL_NETWORK_SINK_PATTERN.search(text)
-    ):
-        patterns.append((_EXFIL_TRANSFER_VERB_PATTERN.pattern, "sensitive transfer intent"))
-    for pattern, reason in patterns:
-        if re.search(pattern, text):
-            signals.append(
-                GuardSignal(
-                    signal_id=f"network:exfil:{reason.replace(' ', '-')}",
-                    family="network",
-                    severity=8,
-                    confidence=0.79,
-                    evidence_source="artifact",
-                    matched_text=reason,
-                    explanation="includes exfiltration-oriented intent",
-                    remediation="Confirm destination and data class before allowing transfer.",
                     rule_version=_RULE_VERSION,
                 )
             )
